some refactoring, plus only using file if csv file already written into gen_funcs directory

+++ libensemble/gen_funcs/__init__.py
@@ -11,24 +11,26 @@ logger = logging.getLogger(__name__)
 class RC:
     """Runtime configuration options."""
 
-    _opt_modules: Optional[Union[str, List[str]]] = None  # optional string or list of strings
-    _is_windows: bool = platform.system() not in ["Linux", "Darwin"]
-    _csv_path = __file__.replace("__init__.py", ".opt_modules.csv")
-
     def __init__(self):
-        if RC._is_windows and os.path.isfile(RC._csv_path):
+
+        self._opt_modules: Optional[Union[str, List[str]]] = None  # optional string or list of strings
+        self._is_windows: bool = platform.system() not in ["Linux", "Darwin"]
+        self._csv_path = __file__.replace("__init__.py", ".opt_modules.csv")
+        self._csv_exists: bool = os.path.isfile(self._csv_path)
+
+        if self._is_windows and self._csv_exists:
             with open(self._csv_path) as f:
                 optreader = csv.reader(f)
-                RC._opt_modules = [opt for opt in optreader][0]  # should only be one row
+                self._opt_modules = [opt for opt in optreader][0]  # should only be one row
 
     @property
     def aposmm_optimizers(self):
         timeout = 5
         start = time.time()
-        if not self._is_windows:
+        if not self._is_windows or not self._csv_exists:
             return self._opt_modules
-        else:
-            while not os.path.isfile(self._csv_path):
+        else:  # is windows and csv exists
+            while not os.path.isfile(self._csv_path):  # self._csv_exists not updated after init
                 time.sleep(0.1)
                 if time.time() - start > timeout:
                     logger.warning("Unable to determine set optimization methods by timeout. Using nlopt as default.")
@@ -48,7 +50,7 @@ class RC:
         if not isinstance(values, list):
             values = [values]
 
-        if self._is_windows and values != current_opt:  # avoid rewriting constantly
+        if self._is_windows and self._csv_exists and values != current_opt:  # avoid rewriting constantly
             with open(self._csv_path, "w") as f:
                 optwriter = csv.writer(f)
                 optwriter.writerow(values)
